update in HydroSourceSMASH::read_in_SMASH_hadrons
There is an option to choose whether the spectators are included
as the source terms.
The corresponding input parameter is Reject_SMASH_spectators.
If non-zero integer value is given, the spectators are not
included and they are listed in a separate file
SMASH_specator.oscar.
The extended SMASH output can be also handled using a parameter
Extended_SMASH_output.