switch neos_bqs back to the Official version
We use a tailored version neos_bqs, which extended to low temperature
and large muB (muB_max = 0.9 GeV) in the Hadron Resonance Gas phase. We
noticed that there is a region (e = 0.2 - 0.4 GeV/fm^3, n_B ~ 0.6 1/fm^3)
where temperature drops to almost 0. We also noticed that the muB has a
sharp transition from 0.6 GeV to 0.9 GeV from the QGP to HRG phase in
the large baryon densition region. Such a sharp transition in muB would
result in artifical large gradients of muB/T in the simulations and
could lead to instability to simulations with finite baryon diffusion.
Therefore, we now switch back to the official neos_bqs version, in which
the muBmax = 0.6 GeV in both phases.