a better treatment of freeze-out corona cells
when there are source terms present, it is hard to find the corona cells
with energy density lower than e_sw. Here we standradize the variable
it_start -> iFreezeStart in the evolve.cpp to identify when the const
tau corona freezeout will be performed. For regular hydrodynamic
simulations without source term, the corona freeze-out will be peformed
at the first time step. When there is source terms present, we first
identify the last source term's tau within mid-rapidity. If it is
smaller than 2 fm/c, we will use it as the starting time for the
freeze-out surface. If it is larger than 2 fm/c, we will start the
freeze-out surface at 2 fm/c to avoid too many fluid cells fall into the
dilute corona region at low energy collisions.

Certainly, the current treatment is not ideal but should be able to take
into account most of the contribution from the cold corona cells. For pp
collisions, this treatment is perfectly good. It becomes worse for
heavy-ion collisions at low energies.