dynamically decide the grid size for InitProf == 113
We first read in the participant list and determine the event center.
We shift the event center to origin. This is useful for asymmetric
collisions such as p+Au collisions.
Then we find the maximum (x, y) coordinate for the participant nucleon
in the event. Finally, the grid size is determined by
  2*(x_max + gridOffset) and 2*(y_max + gridOffset)
Here the gridOffset = min(3, 5*nucleonWidth).
We impose a minimum gridOffset = 3 fm so that when nucleonWidth < 0.6
fm, we still have a large enough grid for the simulations.

We tested central Au+Au collision would have a grid size of [-12, 12]
fm. and p+Au collision would have a grid size of [-7, 7]

In this case, the grid spacing in small collision system will be higher
for better accuracy.