Compatibility with JETSCAPE
Do not use the ${CMAKE_SOURCE_DIR} for the version.h.in file to avoid
confliction when compile MUSIC as an external code package in JETSCAPE
because the variable ${CMAKE_SOURCE_DIR} will be set by the JETSCAPE
directory, not locally from the MUSIC directory.

Simplely omitting it would make the compile work for both JETSCAPE and
MUSIC alone.