update the automatic test for IPGlasma2D
Because we improved the stabilitiy of the eos_hotQCD class in the
previous commit, the IPGlasma2D test failed because the results is no
longer identical to the previous one. We checked that they are still
within negligible difference. Now, we update the numbers in the
IPGlasma2D test for the latest version of the code.