fixed a bug that eps_freeze_max was setting the freeze-out energy density
The bug was introduced on Aug. 31, 2024 (commit #: d597e4d) when I
reformatting the read_in_parameters functions. "eps_switch" needs to be
the last one to modify the epsilon_freeze variable so that it is the
default parameter.

From Aug. 31, 2024 until now, the epsilon_freeze was set to the value of
eps_freeze_max (0.18 GeV/fm^3) and user's input of eps_swtich is not
effective.

Thanks to Xiangyu Wu who pointed out this bug to me!